fix(core): widen module_make_time tv_nsec to i64 for LLP64 targets
The Windows MSVC x86_64 `cargo check` job failed with two E0308
"mismatched types: expected `i64`, found `i32`" errors at
dynamic_module.rs:1541 and :1542.

Root cause: `emacs_time` mirrors C `struct timespec` and must keep its
`#[repr(C)]` ABI field widths:

    pub tv_sec: libc::time_t,   // i64 on all our targets
    pub tv_nsec: libc::c_long,  // i64 on Linux/macOS, i32 on Windows MSVC

`module_make_time` bound `let nanoseconds = time.tv_nsec;` and fed
`nanoseconds / 1000` and `(nanoseconds % 1000) * 1000` straight into
`Value::fixnum`, whose parameter is `i64`. On Linux/macOS `c_long` is
`i64`, so the arithmetic already produced `i64` and compiled; on Windows
MSVC (the LLP64 model, where `long` is 32 bits) the arithmetic stayed
`i32` and could not satisfy `fixnum(i64)`. Linux going green masked the
Windows-only width mismatch.

Fix: widen the field once at the binding, `time.tv_nsec as i64`, so the
downstream arithmetic is target-independently `i64`. The struct and its
ABI repr are deliberately left unchanged -- `tv_nsec` must remain a
`c_long` to match the platform `struct timespec` layout.

Verified: `cargo check -p neovm-core --lib` passes on Linux;
`cargo fmt --all --check` clean.

+++ neovm-core/src/emacs_core/dynamic_module.rs
@@ -1533,8 +1533,13 @@ unsafe extern "C" fn module_make_time(env: *mut emacs_env, time: emacs_time) -> 
         if !module_function_begin(env) {
             return std::ptr::null_mut();
         }
+        // `tv_nsec` is `libc::c_long`, which is `i64` on Linux/macOS but
+        // `i32` on Windows MSVC (LLP64). Widen to `i64` so the arithmetic
+        // matches `Value::fixnum`'s `i64` parameter on every target. The
+        // `#[repr(C)]` `emacs_time` struct itself mirrors `struct timespec`
+        // and must keep its platform-ABI field widths.
         let seconds = time.tv_sec;
-        let nanoseconds = time.tv_nsec;
+        let nanoseconds = time.tv_nsec as i64;
         let list = Value::list(vec![
             Value::fixnum((seconds >> 16) & 0xFFFF),
             Value::fixnum(seconds & 0xFFFF),
